extract fake taxis to separate gateway
