+++ client_dashboard/common/src/commonMain/kotlin/org/thechance/common/domain/getway/ITaxisGateway.kt
@@ -5,13 +5,13 @@ import org.thechance.common.domain.entity.NewTaxiInfo
 import org.thechance.common.domain.entity.Taxi
 import org.thechance.common.domain.entity.TaxiFiltration
 
-interface ITaxiGateway {
+interface ITaxisGateway {
 
     suspend fun getTaxis(
         username: String?,
         taxiFiltration: TaxiFiltration,
         page: Int,
-        numberOfUsers: Int
+        numberOfTaxis: Int
     ): DataWrapper<Taxi>
 
     suspend fun createTaxi(taxi: NewTaxiInfo): Taxi
@@ -19,9 +19,5 @@ interface ITaxiGateway {
     suspend fun updateTaxi(taxi: NewTaxiInfo): Taxi
 
     suspend fun deleteTaxi(taxiId: String): Boolean
-
-    suspend fun searchTaxiByDriverUsername(driverUsername: String): DataWrapper<Taxi>
-
-    suspend fun filterTaxis(): DataWrapper<Taxi>
 
 }
