+++ client_dashboard/desktop/src/jvmMain/kotlin/di/GatewayModule.kt
@@ -5,12 +5,10 @@ import org.koin.core.module.dsl.singleOf
 import org.koin.dsl.module
 import org.thechance.common.data.local.gateway.IdentityGateway
 import org.thechance.common.data.local.gateway.LocalGateway
+import org.thechance.common.data.remote.fakegateway.TaxisFakeGateway
 import org.thechance.common.data.remote.gateway.FakeRemoteGateway
 import org.thechance.common.data.remote.gateway.UsersGateway
-import org.thechance.common.domain.getway.IIdentityGateway
-import org.thechance.common.domain.getway.ILocalGateway
-import org.thechance.common.domain.getway.IRemoteGateway
-import org.thechance.common.domain.getway.IUsersGateway
+import org.thechance.common.domain.getway.*
 
 val GatewayModule = module {
     //TODO remove comment when finish testing
@@ -20,4 +18,6 @@ val GatewayModule = module {
     singleOf(::IdentityGateway) { bind<IIdentityGateway>() }
     singleOf(::UsersGateway) { bind<IUsersGateway>() }
 //    singleOf(::UsersFakeGateway){ bind<IUsersGateway>()}
+    singleOf(::TaxisFakeGateway) { bind<ITaxisGateway>() }
+//    singleOf(::TaxiGateway) { bind<ITaxiGateway>() }
 }
